requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
bitsandbytes==0.49.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
datasets==4.4.1
dill==0.4.0
dotenv==0.9.9
entmax==1.3
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
gitpython==3.1.45
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
jinja2==3.1.6
jsonargparse==3.13.1
lightning-utilities==0.15.2
lxml==6.0.2
markupsafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
networkx==3.6.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pandas==2.3.3
peft==0.18.0
platformdirs==4.5.1
portalocker==3.2.0
propcache==0.4.1
protobuf==4.25.8
psutil==7.1.3
pyarrow==22.0.0
pydantic==2.12.5
pydantic-core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytorch-lightning==2.6.0
pytz==2025.2
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
sacrebleu==2.5.1
safetensors==0.7.0
scipy==1.16.3
sentencepiece==0.2.1
sentry-sdk==2.47.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sympy==1.14.0
tabulate==0.9.0
tokenizers==0.22.1
torch==2.9.1
torchmetrics==0.10.3
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
unbabel-comet==2.2.7
urllib3==2.6.2
wandb==0.23.1
xxhash==3.6.0
yarl==1.22.0
